requirements.txt
# Databricks Apps Requirements
# Python 3.11+ required

# Core framework
streamlit>=1.31.0

# Data manipulation and analysis
pandas>=2.0.0

# Databricks integration
databricks-sdk>=0.20.0
databricks-sql-connector>=3.0.0

# Additional utilities
python-dateutil>=2.8.2
python-dotenv>=1.0.0
